Redesign group screen

+++ app/src/main/java/com/indoor/flowers/fragment/GroupFragment.java
@@ -2,21 +2,29 @@ package com.indoor.flowers.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.design.widget.TabLayout;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import android.support.v4.view.ViewPager;
+import android.support.v4.view.ViewPager.OnPageChangeListener;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageView;
 
 import com.evgeniysharafan.utils.Fragments;
-import com.evgeniysharafan.utils.Toasts;
 import com.indoor.flowers.R;
 import com.indoor.flowers.adapter.EventsAdapter;
 import com.indoor.flowers.adapter.FlowersAdapter;
+import com.indoor.flowers.adapter.FlowersAdapter.FlowersSelectionListener;
+import com.indoor.flowers.adapter.GroupPagerAdapter;
 import com.indoor.flowers.database.provider.DatabaseProvider;
 import com.indoor.flowers.database.provider.FlowersProvider;
 import com.indoor.flowers.database.provider.NotificationsProvider;
@@ -34,27 +42,32 @@ import butterknife.OnClick;
 import butterknife.OnTextChanged;
 import butterknife.Unbinder;
 
-public class GroupFragment extends Fragment implements OnItemClickListener<Notification> {
+public class GroupFragment extends Fragment implements OnItemClickListener<Notification>,
+        OnPageChangeListener, FlowersSelectionListener {
 
     private static final String KEY_GROUP_ID = "key_group_id";
     private static final String KEY_FLOWER_ID = "key_flower_id";
 
     @BindView(R.id.fg_title)
     EditText nameView;
-    @BindView(R.id.fg_events_list)
-    RecyclerView eventsListView;
-    @BindView(R.id.fg_add_event)
-    Button addEventButton;
-    @BindView(R.id.fg_flowers_list)
-    RecyclerView flowersList;
-    @BindView(R.id.fg_delete)
-    Button deleteButton;
+    @BindView(R.id.fg_group_icon)
+    ImageView iconView;
+    @BindView(R.id.fg_tabs)
+    TabLayout tabsView;
+    @BindView(R.id.fg_pager)
+    ViewPager viewPager;
+    @BindView(R.id.fg_add)
+    View addButtonView;
+    @BindView(R.id.toolbar)
+    Toolbar toolbar;
 
     private FlowersProvider flowersProvider;
     private NotificationsProvider notificationsProvider;
     private Unbinder unbinder;
 
     private Group group;
+
+    private GroupPagerAdapter pagerAdapter;
     private FlowersAdapter flowersAdapter;
     private EventsAdapter eventsAdapter;
 
@@ -83,6 +96,8 @@ public class GroupFragment extends Fragment implements OnItemClickListener<Notif
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setHasOptionsMenu(true);
+
         flowersProvider = new FlowersProvider(getActivity());
         notificationsProvider = new NotificationsProvider(getActivity());
         long groupId = getArguments() != null && getArguments().containsKey(KEY_GROUP_ID)
@@ -100,11 +115,10 @@ public class GroupFragment extends Fragment implements OnItemClickListener<Notif
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_group, container, false);
         unbinder = ButterKnife.bind(this, view);
-        initFlowersList();
-        initEventsList();
+        setupActionBar();
+        initPager();
         refreshViewWithGroup();
-        reloadEvents();
-        reloadFlowers();
+        reloadItems();
         return view;
     }
 
@@ -120,36 +134,50 @@ public class GroupFragment extends Fragment implements OnItemClickListener<Notif
         super.onDestroy();
     }
 
-    @OnTextChanged(R.id.fg_title)
-    void onNameTextChanged(CharSequence s, int start, int before, int count) {
-        group.setName(nameView.getText().toString());
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        if (group.getId() != DatabaseProvider.DEFAULT_ID) {
+            inflater.inflate(R.menu.menu_group, menu);
+        } else {
+            super.onCreateOptionsMenu(menu, inflater);
+        }
     }
 
-    @OnClick(R.id.fg_save)
-    void onSaveClick() {
-        if (TextUtils.isEmpty(group.getName())) {
-            Toasts.showLong(R.string.fcr_name_group_error);
-            return;
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.mg_delete) {
+            FlowersAlarmsUtils.deleteAlarmsForEvents(getActivity(),
+                    notificationsProvider.getEventsForTarget(group.getId(), Group.TABLE_NAME));
+            flowersProvider.deleteGroup(group);
+            getActivity().onBackPressed();
         }
 
-        flowersProvider.createOrUpdateGroup(group, flowersAdapter.getSelectedFlowers());
-        refreshDeleteButtonVisibility();
-        refreshEventsVisibility();
-        reloadEvents();
+        return super.onOptionsItemSelected(item);
     }
 
-    @OnClick(R.id.fg_delete)
-    void onDeleteClick() {
-        FlowersAlarmsUtils.deleteAlarmsForEvents(getActivity(),
-                notificationsProvider.getEventsForTarget(group.getId(), Group.TABLE_NAME));
-        flowersProvider.deleteGroup(group);
-        getActivity().onBackPressed();
+    @OnTextChanged(R.id.fg_title)
+    void onNameTextChanged(CharSequence s, int start, int before, int count) {
+        String name = nameView.getText().toString().trim();
+        if (!TextUtils.isEmpty(name) && !name.equals(group.getName())) {
+            group.setName(nameView.getText().toString());
+            flowersProvider.createOrUpdateGroup(group);
+            flowersProvider.refreshGroupFlowers(group, flowersAdapter.getSelectedFlowers());
+            getActivity().invalidateOptionsMenu();
+            refreshAddButtonVisibility();
+        }
     }
 
-    @OnClick(R.id.fg_add_event)
-    void onAddEventClicked() {
+    @OnClick(R.id.fg_add)
+    void onAddClicked() {
+        Fragment fragment = null;
+        if (viewPager.getCurrentItem() == GroupPagerAdapter.POSITION_EVENTS) {
+            fragment = NotificationFragment.newInstance(group.getId(), Group.TABLE_NAME);
+        } else {
+            fragment = FlowerFragment.newInstance();
+        }
+
         Fragments.replace(getFragmentManager(), android.R.id.content,
-                NotificationFragment.newInstance(group.getId(), Group.TABLE_NAME), null, true);
+                fragment, null, true);
     }
 
     @Override
@@ -158,78 +186,111 @@ public class GroupFragment extends Fragment implements OnItemClickListener<Notif
                 NotificationFragment.newInstance(item.getId()), null, true);
     }
 
+    @Override
+    public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
+
+    }
+
+    @Override
+    public void onPageSelected(int position) {
+        reloadItems();
+    }
+
+    @Override
+    public void onPageScrollStateChanged(int state) {
+
+    }
+
+    @Override
+    public void onSelectedFlowersChanged(List<Flower> selectedFlowers) {
+        if (group.getId() != DatabaseProvider.DEFAULT_ID) {
+            flowersProvider.refreshGroupFlowers(group, selectedFlowers);
+        }
+    }
+
+    private void setupActionBar() {
+        AppCompatActivity activity = (AppCompatActivity) getActivity();
+        if (activity != null) {
+            activity.setSupportActionBar(toolbar);
+            ActionBar actionBar = activity.getSupportActionBar();
+            if (actionBar != null) {
+                actionBar.setHomeButtonEnabled(true);
+                actionBar.setDisplayHomeAsUpEnabled(true);
+                actionBar.setTitle(null);
+            }
+        }
+    }
+
     private void refreshViewWithGroup() {
         if (nameView == null) {
             return;
         }
 
         nameView.setText(group.getName());
-        refreshDeleteButtonVisibility();
-        refreshEventsVisibility();
+        refreshAddButtonVisibility();
     }
 
-    private void refreshEventsVisibility() {
-        if (eventsListView == null) {
+    private void refreshAddButtonVisibility() {
+        if (addButtonView == null) {
             return;
         }
 
-        if (group.getId() == DatabaseProvider.DEFAULT_ID) {
-            eventsListView.setVisibility(View.GONE);
-            addEventButton.setVisibility(View.GONE);
-        } else {
-            eventsListView.setVisibility(View.VISIBLE);
-            addEventButton.setVisibility(View.VISIBLE);
-        }
+        addButtonView.setVisibility(group.getId() == DatabaseProvider.DEFAULT_ID ? View.GONE : View.VISIBLE);
     }
 
-    private void refreshDeleteButtonVisibility() {
-        if (deleteButton == null) {
-            return;
+    private void initPager() {
+        initEventsAdapter();
+        initFlowersAdapter();
+        if (pagerAdapter == null) {
+            pagerAdapter = new GroupPagerAdapter();
         }
 
-        deleteButton.setVisibility(group.getId() == DatabaseProvider.DEFAULT_ID ? View.GONE : View.VISIBLE);
+        pagerAdapter.setAdapters(flowersAdapter, eventsAdapter);
+        viewPager.setAdapter(pagerAdapter);
+        viewPager.addOnPageChangeListener(this);
+        tabsView.setupWithViewPager(viewPager);
     }
 
-    private void initEventsList() {
+    private void initEventsAdapter() {
         if (eventsAdapter == null) {
             eventsAdapter = new EventsAdapter();
         }
 
         eventsAdapter.setListener(this);
-        eventsListView.setLayoutManager(new LinearLayoutManager(getActivity()));
-        eventsListView.setAdapter(eventsAdapter);
     }
 
-    private void initFlowersList() {
+    private void initFlowersAdapter() {
         if (flowersAdapter == null) {
             flowersAdapter = new FlowersAdapter();
         }
 
         flowersAdapter.setSelectionMode(true);
-        flowersList.setLayoutManager(new LinearLayoutManager(getActivity()));
-        flowersList.setAdapter(flowersAdapter);
+        flowersAdapter.setSelectionListener(this);
     }
 
-    private void reloadEvents() {
-        if (eventsListView == null) {
+    private void reloadItems() {
+        if (viewPager == null) {
             return;
         }
 
+        if (viewPager.getCurrentItem() == GroupPagerAdapter.POSITION_FLOWERS) {
+            reloadFlowers();
+        } else {
+            reloadEvents();
+        }
+    }
+
+    private void reloadEvents() {
         List<Notification> events = notificationsProvider.getEventsForTarget(group.getId(), Group.TABLE_NAME);
         eventsAdapter.setItems(events);
     }
 
     private void reloadFlowers() {
-        if (flowersList == null) {
-            return;
-        }
-
         List<Flower> allFlowers = flowersProvider.getAllFlowers();
         flowersAdapter.setItems(allFlowers);
 
         List<Flower> flowers = flowersProvider.getFlowersForGroup(group.getId());
         flowersAdapter.setSelectedFlowers(flowers);
-
         flowersAdapter.addSelected(getFlowerIdFromArgs());
     }
